Adds a :realtime option
Adds a feature for realtime test progress output. For large PHPUnit test
suites, the default (non-realtime) implementation buffers all the
PHPUnit ProgressPrinter's output into a string before printing it all at
once. This is because that implementation depends on being able to parse
the output to get the details about the test run (number of tests,
fails, etc)

When :realtime is enabled, a new RealtimeRunner is instead invoked that
has PHPUnit log its test details to a JSON file which can be parsed at
the end of the run. This allows the progress printer to send its output
straight to the terminal.

 - Added LogReader class to parse json ouput
 - Added RealtimeRunner to run PHPUnit with the --log-json flag
 - Changed Runner into a class rather than a module so that
   RealtimeRunner could inherit most of its behavior
 - RealtimeRunner passes a spec list almost identical to Runner's spec
   list